+++ src/dataStructures/bloomFilter/bloomFilter.ts
@@ -1,5 +1,49 @@
+const bitLength = 32;
+
+function hash1(item: string, length: number): number {
+  let hash = 0;
+
+  for (let charIndex = 0; charIndex < item.length; charIndex += 1) {
+    hash = (hash << 5) + hash + item.charCodeAt(charIndex);
+    hash &= hash;
+    hash = Math.abs(hash);
+  }
+
+  return hash % length;
+}
+
+function hash2(item: string, length: number): number {
+  let hash = 5381;
+
+  for (let charIndex = 0; charIndex < item.length; charIndex += 1) {
+    hash = (hash << 5) + hash + item.charCodeAt(charIndex);
+  }
+
+  return Math.abs(hash) % length;
+}
+
+function hash3(item: string, length: number): number {
+  let hash = 0;
+
+  for (let charIndex = 0; charIndex < item.length; charIndex += 1) {
+    hash = (hash << 5) - hash;
+    hash += item.charCodeAt(charIndex);
+    hash &= hash;
+  }
+
+  return Math.abs(hash) % length;
+}
+
+function getHashValues(item: string, length: number): [number, number, number] {
+  return [
+    hash1(item, length),
+    hash2(item, length),
+    hash3(item, length),
+  ];
+}
+
 export class BloomFilter {
-  constructor(size = 100) {
+  constructor(size: number) {
     this.data = Array(size)
       .fill(0);
   }
@@ -7,56 +51,20 @@ export class BloomFilter {
   private readonly data: number[];
 
   public insert(item: string): void {
-    this.getHashValues(item)
+    getHashValues(item, this.data.length * bitLength)
       .forEach((val) => {
-        this.data[val] = 1;
+        const dataIndex = Math.floor(val / bitLength);
+        const bitIndex = val % bitLength;
+        this.data[dataIndex] |= 1 << bitIndex;
       });
   }
 
   public contain(item: string): boolean {
-    return this.getHashValues(item)
-      .every((val) => this.data[val]);
-  }
-
-  public hash1(item: string): number {
-    let hash = 0;
-
-    for (let charIndex = 0; charIndex < item.length; charIndex += 1) {
-      hash = (hash << 5) + hash + item.charCodeAt(charIndex);
-      hash &= hash;
-      hash = Math.abs(hash);
-    }
-
-    return hash % this.data.length;
-  }
-
-  public hash2(item: string): number {
-    let hash = 5381;
-
-    for (let charIndex = 0; charIndex < item.length; charIndex += 1) {
-      hash = (hash << 5) + hash + item.charCodeAt(charIndex);
-    }
-
-    return Math.abs(hash % this.data.length);
-  }
-
-  public hash3(item: string): number {
-    let hash = 0;
-
-    for (let charIndex = 0; charIndex < item.length; charIndex += 1) {
-      hash = (hash << 5) - hash;
-      hash += item.charCodeAt(charIndex);
-      hash &= hash;
-    }
-
-    return Math.abs(hash % this.data.length);
-  }
-
-  public getHashValues(item: string): number[] {
-    return [
-      this.hash1(item),
-      this.hash2(item),
-      this.hash3(item),
-    ];
+    return getHashValues(item, this.data.length * bitLength)
+      .every((val) => {
+        const dataIndex = Math.floor(val / bitLength);
+        const bitIndex = val % bitLength;
+        return Boolean(this.data[dataIndex] & (1 << bitIndex));
+      });
   }
 }